Optimize low level structure operations

+++ include/mata/alphabet.hh
@@ -23,7 +23,7 @@
 #include <mata/ord-vector.hh>
 
 namespace Mata {
-        using Symbol = unsigned long;
+        using Symbol = unsigned;
         using StringToSymbolMap = std::unordered_map<std::string, Symbol>;
 
          /**
